refactor: split monolithic AudioVisualizerDrawable into separate specialized visualizers
